Moved importing of classification evaluation metrics to the top of the analysis.py file.
Added an extra classifier (RandomForestClassifier) to classifiers.py. This also resulted in small changes on line 23 of features.py and line 21 of main.py.

Changed classifiers.py to be lowercase.

Changed the variable "c" to be lowercase in main.py. Changed the input statement to prompt for c and r instead of classification and regression. (Line 16)

Changed line 17 of name.py to work across different machines.