IN MODELS.PY: Added features.featurize and features.datasplit functions to the run() function in order to featurize and split the dataset. Commented out lines 118-120, line 128, line 144, line 147, and line 153, as main.py would not run properly with these lines.
IN NAME.PY: Changed the file directory to match my local directory, as I could not get it to run with any version of "../dataFiles" etc.

IN MAIN.PY: Removed the drop=True argument in line 66, as this is no longer necessary. Commented out line 75, as model.featurization is no longer a function. Removed the tune=True argument in line 77, as this is no longer necessary.